+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Changelog
+
+## 0.2.0
+
+* upgrade rust valopack crate to 1.2.0
+* upgrade FRB to 2.12.0
+* added dart types mirroring for UpdateInfo
+* initialize velopack with URL just once after init
+* added download progress stream sink to Flutter
+* used build hooks instead of cargokit
+* better example UI
+
 ## 0.1.0
 
 * Upgrade Rust bridge to 2.6.0